+++ spec/unit/layer/audio-source.spec.ts
@@ -34,20 +34,20 @@ describe('Unit Tests ->', function () {
 
       it('should update its currentTime when the movie seeks', function () {
         layer.tryAttach(movie)
-        etro.event.publish(movie, 'movie.seek', {})
+        etro.event.publish(movie, 'seek', {})
         expect(layer.currentTime).toBe(2)
       })
 
       it('should update source.currentTime when the movie seeks', function () {
         layer.tryAttach(movie)
-        etro.event.publish(movie, 'movie.seek', {})
+        etro.event.publish(movie, 'seek', {})
         expect(layer.source.currentTime).toBe(layer.currentTime)
       })
 
       it('should update source.currentTime when the movie seeks when sourceStartTime is set', function () {
         layer.sourceStartTime = 0.02
         layer.tryAttach(movie)
-        etro.event.publish(movie, 'movie.seek', {})
+        etro.event.publish(movie, 'seek', {})
         expect(layer.source.currentTime).toBe(layer.currentTime + layer.sourceStartTime)
       })